tested the gradient descent and fixed some bugs. This doesn't work.

+++ NeuralNetworkLibrary/Layer.cs
@@ -46,7 +46,7 @@ namespace NeuralNetworkLibrary
             }
         }
 
-        public void Randomize(Random random, int min, int max) 
+        public void Randomize(Random random, double min, double max) 
         {
             for(int i = 0; i < Neurons.Length; i++)
             {
